requirements.txt
beautifulsoup4==4.4.1
Pillow==3.1.1
selenium==2.52.0
Flask==0.10.1
splinter==0.7.3
